+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 #FileTeCouch CHANGELOG
 
 * August 14, 2014
-	* Release 0.5.0
+	* Release 0.1.0
 	* Initial commit
 	* Basic request : get/set/delete + increment/decrement
 	* Support Couchbase Buckets
 	* Examples
- 
+* Septembre 14, 2014 - 2pm
+	* Release 0.2.0
+	* Add views request
+	* Support queries
+	* Views and queries examples
+* Septembre 14, 2014 - 8pm
+	* Release 0.3.0
+	* Support Memcached Buckets
+	* Examples